backend/requirements.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0

# Database
mysqlclient>=2.2,<3.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0

# Environment variables
python-dotenv>=1.0,<2.0

# Email
django-anymail>=10.2,<11.0

# Utilities
python-dateutil>=2.8,<3.0
uuid>=1.30

# Development
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0
factory-boy>=3.3,<4.0

# Production
gunicorn>=21.2,<22.0
whitenoise>=6.6,<7.0
